+++ doc/user/content/release-notes.md
@@ -107,6 +107,9 @@ List new features before bug fixes.
 
 - Prevent overflow on operations combining [`timestamp`] and [`interval`].
 
+- Support explicit reference to the default collation (`<expr> COLLATE pg_catalog.default`).
+  Other collations are not yet supported.
+
 {{% version-header v0.10.0 %}}
 
 - Allow ingesting avro schemas whose top node is not a record type.